Installed tailwind and added styles
