+++ client/src/components/Posts.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { deletePost, fetchPosts } from '../redux/posts/postsSlice';
+import { deletePost, fetchPosts } from '../redux/postsFunc';
 import { Link, useNavigate } from 'react-router-dom';
 
 function Posts() {
@@ -22,20 +22,25 @@ function Posts() {
   };
 
   return (
-    <div>
+    <section className="max-w-4xl mx-auto my-5 flex gap-x-5 gap-y-5 justify-center flex-wrap">
       {posts.map((post) => (
-        <div key={post.id}>
+        <article
+          key={post.id}
+          className="flex flex-col gap-y-4 border border-slate-900 rounded bg-slate-100 p-2 w-72"
+        >
           <Link to={`/posts/${post.id}`}>
-            <h2>{post.title}</h2>
-            <p>{post.body}</p>
+            <h2 className="font-medium text-2xl mb-1">{post.title}</h2>
+            <p className="text-md">{post.body}</p>
           </Link>
-          <Link to={`/posts/${post.id}/edit`}>Edit</Link>
-          <button type="button" onClick={() => handleDelete(post.id)}>
-            Delete
-          </button>
-        </div>
+          <div className="flex justify-between">
+            <Link to={`/posts/${post.id}/edit`} className='bg-green-600 py-1 rounded w-16 text-center text-white font-semibold'>Edit</Link>
+            <button type="button" onClick={() => handleDelete(post.id)} className='bg-red-600 py-1 rounded w-16 text-center text-white font-semibold'>
+              Delete
+            </button>
+          </div>
+        </article>
       ))}
-    </div>
+    </section>
   );
 }
