+++ client/src/index.css
@@ -1,4 +1,13 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 body {
   margin: 0;
   padding: 0;
+}
+
+.active {
+  text-decoration: underline;
+  text-decoration-style: wavy;
 }
